backend/requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
openai